fix: 책임 분리를 위해 컨트롤러 추가 및 수정
